requirements.txt
yt-dlp
flask
discord.py
python-dotenv